Extract the element-pool solver into its own module
The chemical-potential solve had grown into ~320 lines of stateless
numerical code inside defect_system.py: occupancy statistics over
per-group state arrays, the staged Newton solves, the bracketing
sweep, and the target verification, none of which touch DefectSystem
state. These move to py_sc_fermi.element_pools as module-level
functions over a GroupData representation (n_free, log Boltzmann
weights, stoichiometry vectors per group). DefectSystem keeps the
domain side -- species and pool resolution, group assembly -- and
delegates through a thin _solve_chemical_potentials wrapper.
ElementPoolError lives in the new module and is re-exported from
defect_system. Pure move; no behaviour change.
